requirements.txt
openai
pandas
numpy
langchain
langchain-openai
langgraph
chromadb
pydantic
fastapi
uvicorn
python-dotenv
langchain-core
langchain-community
pypdf